moved and/or
