+++ stan/math/prim/scal/meta/contains_vector.hpp
@@ -2,7 +2,7 @@
 #define STAN_MATH_PRIM_SCAL_META_CONTAINS_VECTOR_HPP
 
 #include <stan/math/prim/scal/meta/is_vector.hpp>
-#include <stan/math/prim/arr/meta/or.hpp>
+#include <stan/math/prim/scal/meta/or.hpp>
 
 namespace stan {
